link_tracker_qr_code/requirements.txt
qrcode==7.4.2
Pillow==9.4.0
